api/requirements.txt
cryptography==45.0.5
Django==4.2.23
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
pillow==11.1.0
simplejson==3.20.1
ollama==0.4.7
# sentence-transformers==4.1.0
python-dotenv==1.1.0
psycopg[binary,pool]==3.2.9
pgvector==0.4.1